[10.0][MIG] project_stage_state: Upgrade to v10
