+++ project_stage_state/i18n/fi.po
@@ -8,53 +8,53 @@ msgid ""
 msgstr ""
 "Project-Id-Version: Odoo Server 10.0\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2017-03-17 09:31+0000\n"
-"PO-Revision-Date: 2017-03-17 09:31+0000\n"
+"POT-Creation-Date: 2017-08-10 02:49+0000\n"
+"PO-Revision-Date: 2017-08-10 02:49+0000\n"
 "Last-Translator: OCA Transbot <transbot@odoo-community.org>, 2017\n"
-"Language-Team: Slovenian (https://www.transifex.com/oca/teams/23907/sl/)\n"
+"Language-Team: Finnish (https://www.transifex.com/oca/teams/23907/fi/)\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: \n"
-"Language: sl\n"
-"Plural-Forms: nplurals=4; plural=(n%100==1 ? 0 : n%100==2 ? 1 : n%100==3 || n%100==4 ? 2 : 3);\n"
+"Language: fi\n"
+"Plural-Forms: nplurals=2; plural=(n != 1);\n"
 
 #. module: project_stage_state
 #: selection:project.task.type,state:0
 msgid "Cancelled"
-msgstr "Preklicano"
+msgstr "Peruutettu"
 
 #. module: project_stage_state
 #: selection:project.task.type,state:0
 msgid "Done"
-msgstr "Opravljeno"
+msgstr "Valmis"
 
 #. module: project_stage_state
 #: selection:project.task.type,state:0
 msgid "In Progress"
-msgstr "V teku"
+msgstr ""
 
 #. module: project_stage_state
 #: selection:project.task.type,state:0
 msgid "New"
-msgstr "Novo"
+msgstr ""
 
 #. module: project_stage_state
 #: selection:project.task.type,state:0
 msgid "Pending"
-msgstr "Na čakanju"
+msgstr "Odottaa"
 
 #. module: project_stage_state
 #: model:ir.model.fields,field_description:project_stage_state.field_project_task_state
 #: model:ir.model.fields,field_description:project_stage_state.field_project_task_type_state
 msgid "State"
-msgstr "Stanje"
+msgstr ""
 
 #. module: project_stage_state
 #: model:ir.model,name:project_stage_state.model_project_task
 msgid "Task"
-msgstr "Opravilo"
+msgstr "Tehtävä"
 
 #. module: project_stage_state
 #: model:ir.model,name:project_stage_state.model_project_task_type
 msgid "Task Stage"
-msgstr "Stopnja opravila"
+msgstr "Tehtävän vaihe"
