+++ project_stage_state/models/project_task.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+# Daniel Reis, 2014
+# GNU Affero General Public License <http://www.gnu.org/licenses/>
+
+from odoo import models, fields
+
+
+_TASK_STATE = [
+    ('draft', 'New'),
+    ('open', 'In Progress'),
+    ('pending', 'Pending'),
+    ('done', 'Done'),
+    ('cancelled', 'Cancelled')]
+
+
+class ProjectTaskType(models.Model):
+    _inherit = 'project.task.type'
+    state = fields.Selection(_TASK_STATE, 'State')
